Move all selectors to a single place
