+++ app/src/components/TaskTitle/index.tsx
@@ -2,7 +2,8 @@ import React from 'react';
 import Emoji from 'components/Emoji';
 import Stack from 'components/Stack';
 import styles from './styles.module.css';
-import { updateTaskTitle, getTitle } from 'modules/task';
+import { updateTaskTitle } from 'modules/task';
+import { getTitle } from 'modules/selectors';
 import Input from 'components/Input';
 import useSelector from 'hooks/useSelector';
 import useDispatch from 'hooks/useDispatch';
