+++ app/src/components/TaskTodos/index.tsx
@@ -9,8 +9,9 @@ import styles from './styles.module.css';
 import useDispatch from 'hooks/useDispatch';
 import useSelector from 'hooks/useSelector';
 
+import { getTodos } from 'modules/selectors';
+
 import {
-  getTodos,
   moveTodo,
   newTodo,
   pasteTasks,
